+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Improved the command line handler to support arrow keys and history using up key
 - Log rotators that take longer to archive now correctly resolve subsequent format specifiers
 - Log rotator did not resolve correct `%level` for destination directory
+- Key is reset if client connected after client is dead before removal
 
 ### Updates
 - Debug printed in server version in connection response if build has debug logs enabled